% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ef_addcomma}
\alias{ef_addcomma}
\title{Add comma to number and round to digits}
\usage{
ef_addcomma(x, dig = 0)
}
\arguments{
\item{x}{any numerical value}

\item{dig}{round digits}
}
\value{

}
\description{
Add comma to number and round to digits
}
\examples{
\dontrun{

ef_addcomma_zr(100000.12244)

}
}
