% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ef_conv_perc}
\alias{ef_conv_perc}
\title{Convert number to text and percentage}
\usage{
ef_conv_perc(x)
}
\arguments{
\item{x}{number to convert to percentage}
}
\value{

}
\description{
Convert number to text and percentage
}
\examples{
\dontrun{

ef_conv_perc(.1234)

}
}
