% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ef_create_datemap}
\alias{ef_create_datemap}
\title{Create a datemap with columns}
\usage{
ef_create_datemap()
}
\value{

}
\description{
Create a datemap with columns
}
\examples{
\dontrun{

ef_create_datemap()

}
}
