% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ef_date_diff}
\alias{ef_date_diff}
\title{Number of business days or all days datediff}
\usage{
ef_date_diff(start, end, bus_only = FALSE)
}
\arguments{
\item{start}{start date of filter}

\item{end}{end date of filter}

\item{bus_only}{indicate business only}
}
\value{

}
\description{
Number of business days or all days datediff
}
\examples{
\dontrun{

ef_date_diff('2020-01-01','2020-12-31',bus_only = FALSE)
ef_date_diff('2020-01-01','2018-12-31',bus_only = FALSE)
ef_date_diff('2020-01-05','2020-01-05',bus_only = FALSE)
ef_date_diff('2020-01-01','2020-12-31',bus_only = TRUE)

}
}
