% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ef_if.na}
\alias{ef_if.na}
\title{If NA, NAN, or Infinate repalce with value}
\usage{
ef_if.na(x, y)
}
\arguments{
\item{x}{value to be checked}

\item{y}{value to replace if x is NA, NAN, or Infinite}
}
\value{
a data frame of cots symbols with descriptions
}
\description{
Uses ifelse to replace NA only if value is NA
}
\examples{

\dontrun{

ef_if.na(NA,3)
ef_if.na(4,5)

}
}
