% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ef_is.nainf}
\alias{ef_is.nainf}
\title{Is NA, NAN, or Infinite}
\usage{
ef_is.nainf(x)
}
\arguments{
\item{x}{value to be checked}
}
\value{

}
\description{
Is NA, NAN, or Infinite
}
\examples{
\dontrun{

ef_is.nainf(NA)
ef_is.nainf(1)

}
}
