% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ef_overlap}
\alias{ef_overlap}
\title{Check Vector Overlap}
\usage{
ef_overlap(vec1, vec2, in_out = "Out")
}
\arguments{
\item{vec1}{first vector}

\item{vec2}{second vector}

\item{in_out}{Out will return non matching and In will return matching}
}
\value{

}
\description{
Will return overlapping or non overlapping variables
}
\examples{
\dontrun{

ef_overlap(c(1:10),c(7:15),'Out')
ef_overlap(c(1:10),c(7:15),'In')

}
}
