% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ef_tibcol}
\alias{ef_tibcol}
\title{Replace spaces with . in column header and convert to tibble}
\usage{
ef_tibcol(tibDF)
}
\arguments{
\item{tibDF}{Tibble or Dataframe}
}
\value{

}
\description{
Replace spaces with . in column header and convert to tibble
}
\examples{
\dontrun{

testdf = dplyr::tibble(`test col` = c(1:10))
ef_tibcol(testdf)

}
}
